Now downloads LiveConnect Reports for improved speed and compatibility
This script update now downloads the majority of reports with the LiveConnect option for improved speed and compatibility. This uses the REST API in PowerShell instead of using the limited PowerShell cmdlet.

This change also solves for a recently discovered issue:  

The PowerShell cmdlet downloads Reports as-is. If they include the model, the model comes with it. There is no option to choose LiveConnect. If a Report/Model is set in the 'Large Semantic Model' format, the report cannot be downloaded via API unless it's in LiveConnect mode. The script would get an error and skip that report when attempting to download.

Downloading all Reports via LiveConnect with the REST API solves this limitation and also speeds up the script.

Note: for Pro and Non-Premium workspaces, the full Model's Report will continue to be downloaded to extract the Model from the PBIX. This means that any Pro workspace's Report/Model cannot be in Large Model format.

A timer for total elapsed time was also added.