Unused Model Object detail is now included in final output model template
This includes updated updates to apply Flags and detail on where and how a Model object is used and....more importantly, NOT USED across Reports, Measures, and Relationships. 

Unused Model Objects:

Objects within a Model that are not used in any part of a Report, the building of a Measure, or part of a Relationship.